Extinct & showborn
This is a combination of "extinct" and "showborn" patches which were
separate for earlier versions.

The extinct patch makes extincted species listed among the genocide
list (marked with the "(extinct)" tag) when the game ends.

Notice that the counters used to check extinction keep track of the
number of each monster CREATED, not ones killed. So, if a monster is
marked extinct, it just means that none of them would have been
created any more in the game, not necessarily that you killed them
all. Some may still be alive somewhere in the dungeons.

The "showborn" patch will optionally add the number of each monster
species created after the number of them killed, if the number is
different. No born counter is shown for species with no killed
individuals, or for those species with all born individuals killed.

It also adds the wizard mode extended command #showkills that lets
you see the kill counters of each monster race as you see them when
the game ends.

Download and apply the patch and add the definition
somewhere in your include/config.h to compile the features in.
You also need to turn the "showborn" option on to actually show the
born counters. Use the 'O' option in the game or edit your
configuration file.

Signed-off-by: Wingnut
Signed-off-by: Jukka Lahtinen
Signed-off-by: Steven Noonan <steven@uplinklabs.net>