Add a new configuration option MSGTYPE. This allows the user to configure whether a message always waits for --more--, or if it's never shown, or if it's never shown more than once in row.
The format is
MSGTYPE=foo "string to match"
Where "foo" is one of "hide", "stop" or "norep", and the string to match
is a message to be matched, with '*' and '?' wildcard expansion.

"hide" hides the message, so it's never shown.
"stop" always prompts for --more-- after the message.
"norep" shows the message, but not again if no other message is shown in between.

("noshow" is an alias to "hide", and "more" is an alias to "stop")


For example:

# displacing pets
MSGTYPE=hide  "You displaced *"
# items on floor
MSGTYPE=norep "You see here a *"
MSGTYPE=norep "You see here an *"
# delayed instadeaths
MSGTYPE=stop  "You are slowing down."
MSGTYPE=stop  "You find it hard to breathe."
MSGTYPE=stop  "You are turning a little *"




git-svn-id: svn://sporkhack.com/nh@816 0b9a868d-f130-0410-bc2a-adc7ae8015a6