CS38: Level: new branch "Town"
The Town is an upward dungeon branch whose entrance is a little deeper
than Sokoban. It consists of 4 normal filler levels, capped off by 1 of
2 randomly chosen town maps, inspired by Angband and having several
stores, a new mugger enemy and the Watch.

I like the idea of this branch, but I think it's boring in its current
form with the 4 filler levels. KoboldLord's LP of UnNetHack has a
relevant quote (before he first hit the Town):

http://forum.rpg.net/showthread.php?466187-(Let-s-Play!)-Sporkhack&p=11465246#post11465246

> Sooo... Four levels of utterly nondescript dungeon, all without
> anything that gives me motivation to ever come this way again.

If this were me, I'd reduce this branch to 1 filler level, using maze
generation, filled with leprechauns, nymphs and squeaky board traps.