CS75: Pasi Kallinen's new level compiler
This took a LONG time and a lot of effort. Phew!

That said, there are some things to note between this commit and
UnNetHack's (usual conversion stuff aside):

 * The changes to src/quest.c are omitted. They look unrelated and don't
   feature in paxed's original patch.

 * The original changeset has a minor bug (which I've fixed) in
   sp_level_coder in src/sp_lev.c, where n_rmon in case
   SPO_RANDOM_MONSTERS is checked for out-of-boundsness, but it uses &&
   instead of || like SPO_RANDOM_OBJECTS and SPO_RANDOM_PLACES above it.

 * add_opcode in util/lev_main alloc's and memcpy's for every added
   opcode, which is an awful O(n^2) implementation. The right way to do
   this would be to use realloc; I haven't here because in practice the
   level compiler is stupidly fast, runs only when compiling the game
   and would take millions of runs to recoup the time to write and debug
   new code, so it stays.

The biggest caveat about this commit is that random places may occur in
subroom walls, which can be seen in Minetown or the top of the Town
branch: monsters can spawn in walls, and the stairs down at the top of
Town may be placed in a wall, making it impossible to leave without a
means of digging or level teleport (it has no traps). This is a known
problem with this patch (see the new README), so UnNetHack must have
found a way around it in future changesets.