Don't output '-e' in launch script on Ubuntu
This suppresses an error where the launch script contains an extra '-e'
when built on Ubuntu.

The CMake file spits out 'echo -e ...' to the shell script generating
makefile, which runs under /bin/sh. /bin/sh links to dash, not bash, on
Ubuntu, and dash's echo doesn't accept '-e' as an option.

In short: echo is not portable between shells, but printf is.