CS100: Plane of Air: add up/dn telergns, note bug
I kept the regions out of CS10, but I realize that's a mistake, since it
makes teleporting on the portal side of the Plane of Air send you back
to the far side of the clouds. For now, we'll adopt the same bug as
UnNetHack at this point in its history.