CS45,57,91,141,213: Shuffle dragon appearances
BUG: Appearances (name, colors) should be shuffled, but only name does:
colors still map to their original magical effects, which is wrong.

CS45 contains the dragon appearance shuffling. CS57, CS91 and CS141 fix
bugs with gold dragons due to this change. CS213 includes baby dragons
in the shuffle. See these on Trac by putting the number at the end of
this URL:

    http://sourceforge.net/apps/trac/unnethack/changeset/

The shuffle (should) work as follows:

 1. If starting a new game, shuffle the dragon armors in objects[] at
    the end of shuffle_all().

 2. After that, or after loading objects[] from a restored game, move
    the dragon names and colors in mons[] to match the shuffled armors.
    This is done by calling the new dragons_init() function in
    post_init_tasks(). This requires mons[] to be made un-const again.

The only real change between this and the original changeset is that the
pristine dragon order in mons[] is preserved in a static array in
dragons_init(), since the NitroHack process may play more than one game,
de-syncing it from the order implied by oc_name_idx.