CS116-117: New Nymph level from SLASH'EM
This introduces Aphrodite, a purple 'n' with 10/15/-5 hit dice, speed
and AC respectively, but is otherwise like other nymphs. Her tile can be
found in UnNetHack changeset 116.

The Nymph level itself has Aphrodite, occurs in 45% of games on a level
between 5 and 19 inclusive, and is a normal rooms-and-corridor level
with guaranteed squeaky board traps in each room and only generates
nymphs. The large room with Aphrodite has 4 chests, 4 rings, 5 gems,
4 other random items and 6 more nymphs.

If I'm right about monster generation, it should be completely safe to
drop all of your items on another level and punch all the nymphs to
death, but hitting Aphrodite like that might be tough.