CS118: Patch: Dungeon Growths by Pasi Kallinen
Original patch: http://bilious.alt.org/?35

This patch has a number of related effects:

 * Trees may grow near other trees. The way this is done will never
   block access to any part of the dungeon.

 * Trees may drop fruits and eucalyptus leaves.

 * Herbs on the ground and kelp fronds in water may grow in number if
   left for long enough. Water currents may also move objects around in
   them.

These may happen even if you leave and return to the level. This seems
like it has potential for farming, but even under extreme circumstances,
the amount of food spent waiting for all this dwarfs the amount of food
gained back, even with a ring of slow digestion. The only real effect is
to make the dungeon feel a bit more alive.

Took me a while to find the right place to put the call to
catchup_dgn_growths(), since it didn't belong in getlev() in restore.c
due to NitroHack storing all levels in memory. Also note the use of
moves and level->lastmoves compared to the original code in UnNetHack.

All these new functions take a (struct level *) as an argument, even
though not all the functions and macros they call take advantage of
that. My policy (past and future) on adding functions is to always add
this argument if its needed, even if the level global is available.
True, level is used practically everywhere, but not having a
(struct level *) creates nasty segfault surprises when a function
suddenly finds itself being called while a level is being made. Using
the argument sidesteps that and generalises better.