CS122,126,230: Nerf scrolls of genocide
Uncursed scrolls of genocide kill all of a single monster on the current
level, while blessed ones erase them from the whole dungeon. It is no
longer possible to genocide a whole class of monsters with a single
genocide.

Changeset 126 makes 122 compilable. Changeset 230 fixes thrones to give
dungeon-wide genocides.

I have mixed feelings about this change. It's in the same category of
nerfs that just make it harder to do the same thing as in vanilla
NetHack, but it also aligns closer to how the scrolls are actually
used, e.g. genociding sea creatures on the Plane of Water. It also means
that players must deal with liches and mind flayers instead of creating
easy mode for their late games.

I realise the irony of not giving kill_monster_on_level() a
(struct level *) argument right after the last commit, but it's clearly
modelled after kill_genocided_monsters() above it, and that doesn't have
such an argument either, so I thought it best to stay consistent than
blindly pursue ideals.