CS133: Cursed thrown objects may hit the ceiling
... and fall back on your head. Specifically, it occurs 1/3 of the time
the object would normally slip to the floor under you. In practice, this
means it only happens after a 1/7 check (chance of a cursed throwing
object slipping), then a 1/9 check (chance of the slip not flying in an
x/y direction), then finally the 1/3 check just described: a probability
of 1/189; a rare chance indeed! Play-testing had me throw cursed darts
nearly 200 times before it finally happened.