CS139: Update data.base with entries from SLASH'EM
Compared to the original changeset, justified text was unjustified,
periods were double-spaced, and all text has been wrapped to fit in 72
columns or less (including the leading tab). The pyrolisk entry was
added to the existing one instead of overriding it, and the cockatrice
entry was changed so it wouldn't override the chickatrice entry.

On a general note, the database is a feature that could have been useful
but isn't. Imagine a database that gives newbies basic strategy, and
gives veterans a quick reference for comparing weapons, armor and
looking up what intrinsics monsters have or grant. This feels like a
product of the NetHack DevTeam's no-spoilers policy, which likely used
to enhance the mystery and social aspect of the game i.e. sharing tips,
but nowadays just pushes players out of the game in order to succeed
within it.