CS140,151: Add Cthulhu from SLASH'EM Lethe
Cthulhu replaces the high priest of Moloch as the holder of the Amulet
of Yendor. Don't be fooled by his priestly appendage; Cthulhu is a very
tough fighter, sucks brains like a mind flayer, grabs, possesses a
confusing gaze, has Demogorgon-like hit dice, high speed and armor
class. When killed he leaves a stinking cloud, and he revives when the
cloud dissipates. He is generated with a wand of death and a potion of
full healing.

Cthulhu is pretty hardcore: in play-testing with an average ascension-
level character (melee type), I barely managed to kill him from full
health with a few potions of full healing. It is possible, though.

When I ported this, I misspelled Cthulhu in extern.h, leading to me
having to recompile everything. It'd be nice if the prototypes in
extern.h were split out and used piecemeal by the files that need them
to avoid that.