Record Elemental Planes consistently in history
This records entering just the first Elemental Plane entered and the
Astral Plane, instead of having some recorded and some not due to the
new randomized order. This complements the Randomized Planes commit
(CS10, 1e2505febf).

This also correctly reports the names of the Elemental Planes properly
when you e.g. kill the Wizard of Yendor on one of them.

Upward-facing branches in general aren't tracked upwards, so it's
impossible to track how far you make it into them since the code records
the entry level as the deepest level for that dungeon and ignores
further upward progress. The dunlev_reached tracking could be changed,
but it would need a lot of testing to make sure subtle things elsewhere
don't break.