CS152: Spice up Moloch's Sanctum
The exterior of the level is now lava, fenced off with windows of iron
bars. The shape of the level stays the same.

Two drawbridges connect the two halves of the level, so you'll need a
way to open them (e.g. knock, wand of opening) or break them down and
cross lava (e.g. force bolt, wands of striking and cold).

The entrance half of the level has a player monster trapped in a small
cage of iron bars, complete with fake plastic imitation of the Amulet
of Yendor.

The chamber surrounding the temple now has two mind flayers that carry
wands of striking. A secret door stops them from reaching the
drawbridges, but if they follow you out, they might break them.