CS163: labelread patch (by L)
Original patch: http://bilious.alt.org/?202

Allows the following non-scroll objects to be read:

 * gold pieces
 * tins
 * cans of grease
 * candy bars
 * credit cards
 * magic markers
 * the Orb of Fate
 * runeswords, runed broadswords and daggers

Most of these won't be listed in the read prompt, but you can still
choose them with their inventory letters. Tins, cans of grease and candy
bars have no labels, but all the others give flavor messages. Try being
an elf or archaeologist for different messages for the last items in
that list.

UnNetHack does not include the coin flipping patch that the original
patch rolls in, so it's not here either.