CS178,179,195,288,301: Tinfoil hat patch (by L)
Original patch: http://bilious.alt.org/?85

The tinfoil hat is head armor that blocks telepathy and clairvoyance,
and stops damage from the psi-bolt monster spell and mind flayer psychic
blasts. It has a 90% chance to be generated cursed.

Changeset 178 contains the bulk of the tinfoil hat code. 179 fixes
tinfoil hats blocking intrinsic telepathy acquisition. 195 makes tinfoil
hats non-magical. 288 fixes amulets of ESP broken by 178. 301 prevents
impossible() calls when wearing or taking off tinfoil hats.

Compared to the UnNetHack changes, do.c uses the BTelepat macro which is
identical to the check it was making, and polyself.c has a fix for a
logic error in the multi-line condition. Finally, which_armor() had to
be changed to take a const monster pointer due to some warnings.

The tinfoil hat has a tiles graphic, which can be found in changeset 178
of UnNetHack.