CS251: Bag of Tricks Enhancement patch by Nephi
Original patch: http://bilious.alt.org/?58

A bag of tricks does something interesting about 1 in 4 times it is
applied:

 * 2/9 spit out a random item (of weight >= 100)
 * 1/9 bag wriggles away from you
 * 1/9 bag tries to pull you in, paralyzing you for d4 turns
 * 1/9 a stinking cloud erupts from the bag
 * 1/9 hallucinates you for d35 + 10 turns
 * 1/9 yells "Boo!" and scares monsters in sight (sometimes you too)
 * 1/9 grows teeth and bites, like when looted
 * 1/9 licks your wielded weapon out of your hand

This patch also adds YAFM for the bite when hallucinating, and lets you
use the bag like a sack when its out of charges. Watch out, charging a
bag of tricks destroys its contents!