CS308,410: lev comp: braces,RANDLINE,MAZEWALK loot
This commit changes the following about the special level compiler:

 * Curly braces for rooms, subrooms and IF conditions
 * New RANDLINE command: draws a line of terrain between two coords with
   a given thickness and roughness
 * MAZEWALK can optionally omit loot generation
 * Allow whitespace before comment on a line

This change also creates slightly less lava spills and adds random
rivers to the Gehennom fill levels.

This also fixes several levels where monsters appeared in walls.
Changeset 410 fixes this for one of the Town variants that was forgotten
in 308.

Changeset 308 of UnNetHack forgets a semicolon for randline_detail in
lev_comp.y which is fixed in this commit.

Changes originally from SporkHack r466-475 and r477.