CS334: !GA tries 3 times; blessed raises just once
An uncursed potion of gain ability tries 3 times (instead of 6) to boost
an attribute, whilst a blessed potion tries 1000 times but only increases
one attribute (instead of all of them). This nerfs potions of gain
ability.

I ran the numbers: the chance of a blessed potion failing to find an
attribute to increase is so miniscule that it's somewhere in the order
of 1 / 10^-80. I don't get why the code doesn't just guarantee an
attribute boost for blessed potions.