CS363: Nerf scrolls of earth in Black Market
In the Black Market, each boulder from a scroll of earth has only a 50%
chance of being created.

I get why this was done, but this solution is arbitrary and corner-
casey. UnNetHack changeset 942 lets One-eyed Sam smash boulders when
angry; combine this with, say, a wand of striking or two on his
assistants, and suddenly there's no reason to nerf scrolls of earth in
this one level.