CS368,369,551,552: Tutorial Mode by Alex Smith
This adds a new tutorial mode option to the title screen. Starting a
game in this mode gives players help messages, guiding them through the
basics of the game, based on what they see and do.

Whew, I had to add and change a lot of code to make this work. Compared
to UnNetHack:

 * Tutorial mode is entered by a title screen option, not the character
   selection prompt. This adds front-end logic to display and select
   tutorial character options, so future clients will need to add such
   logic too.

 * AFAIK, NitroHack doesn't track saves or restores, so tutorial code
   that tries to stop too many tutorial messages appearing after
   restoring is removed, otherwise it could make games unreplayable.

 * %Cm is added for menu-like text display in quest.txt, but NitroHack
   already does this for %Cc (cover screen, %Cp is pline) so no changes
   were needed to the delivery method in com_pager (pager.c).

 * flags.numpad no longer exists, so this shows numpad messages by
   default and displays the matching vi-keys message in the tutorial
   review menu.

 * Wizard mode level teleport and the tutorial review both use the same
   command callback and are thus mutually exclusive. The modes
   themselves are too, so this isn't a problem for now. Ctrl-V is moved
   from the levelteleport extended command to tutorialreview.

 * The code to view things in tutorial.c was altered with
   nh_describe_pos of pager.c as the example to copy. Note also the
   explicit init, saving and restoring of static tutorial state, the
   translation of NitroHack commands to classic char codes for the
   tutorial command buffer, and the removal of code for the scroll of
   mail (SCR_MAIL, it doesn't exist in NitroHack, but it'd be easy to
   add again).

 * UnNetHack forgets to add new functions to extern.h, which I've fixed
   here.

 * Client-server games haven't been tested, but they should work.

There's room for improvement. This doesn't change the tutorial text at
all, so there are likely references to "NetHack". The text itself could
stand to use a lot of editing, so they're not walls of text. The
messages could be spaced apart a bit more; right now, it feels like an
info dump. It's tricky to find recently-shown messages in a long
tutorial review menu.

Changeset 368 contains the bulk of the tutorial mode. 369 fixes a small
bug in the tutorial review menu during UnNetHack's development. 551 adds
help when choosing an item class in the partial inventory, and 552 adds
help when first prompted for an item.