Show item tutorial msg when picking item actions
This assumes that the menu of item actions is assembled as the player
asks for it, since it shows in the function that builds it.

I wanted it to go where the menu is shown, but apparently that lives in
the front-end interface, so this will have to do for now.