CS374: Fix non-existing potions in color alchemy
If two potions mixed with color alchemy produced a non-existing output
potion, it would produce that instead of one of the random alchemy
results.

Compared to UnNetHack, this omits some defensive programming put in
places that will almost certainly annoy the player, e.g. a warning in
the xname2 function.