Fix nerf of scroll of earth in Black Market
I made a mistake porting UnNetHack changeset 363 with the display of the
message "But nothing else happens" where a boulder dropped on the player
but nowhere else would wrongly show that message.

Extracted from changeset 393, monsters who read a scroll of earth in the
Black Market now have just a 50% chance of each boulder appearing, just
like for the player.

These changes are meant for commit e12464b40d.