Conduct: Thievery checks Cthulhu, not High Priest
The master thief conduct is to take the artifacts needed to ascend
without killing their holders. Since Cthulhu now holds the Amulet of
Yendor, he should be checked for this in is_guardian() rather than the
High Priest of Moloch.

This bug also exists in UnNetHack. This is meant for commit d9755fc6a1
(extended conducts).