CS409: Instrument shops (NetHack Brass)
The top level of Town has a 2 in 3 chance of having an instrument shop.
The chances for items to appear in a music store are:

 * 10% tin whistle
 *  3% magic whistle

 * 10% wooden flute
 *  3% magic flute

 * 10% tooled horn
 *  3% frost horn
 *  3% fire horn
 *  3% horn of plenty

 * 10% wooden harp
 *  3% magic harp

 * 10% bell
 * 10% bugle

 * 10% leather drum
 *  2% drum of earthquake

 *  5% T-shirt
 *  5% lock pick

Compared to UnNetHack, this commit reintroduces init_shop_selection()
without renaming it to shop_selection_init(), since the other _init()
functions aren't called any more.