CS433: Remove Guidebook.txt
Guidebook.txt should be the output of Guidebook.tex (or is it Guidebook.mn),
though neither NitroHack nor UnNetHack actually generate it as a build step
(correct me if I'm wrong on any of this).

I myself would like to see the Guidebook or something like it included
with the game itself so it can be read within the game, split into
sections that can be selected from menus.