CS436: Add new options to Guidebook.tex
This adds pickup_dropped and pickup_thrown to the guide book TeX file.

Compared to the UnNetHack changeset, this commit adds pickup_thrown (not
included with NitroHack's Guidebook.tex yet), doesn't add pilesize
(omitted since it was redundant in NitroHack), and doesn't make
corrections to dumpfile, newcolors or videocolors (options that don't
exist in NitroHack).