CS441: Enable standout option by default
This highlights the --More-- that appears when multiple messages are
shown in a single turn by default. Some extra trailing whitespace in the
code is also stripped from the sidebar option line.

Compared to the UnNetHack changeset, this doesn't change the msg_window
option, since it doesn't exist in NitroHack, which always shows a list
of messages.