CS445: lev comp: OBJ/MON args, LOOP+SWITCH (Spork)
This adds the SWITCH/CASE/BREAK structure to the level compiler.  SWITCH
takes an integer n in square brackets and matches it to the CASE with
rn2(n), so CASE numbers need to be 0 to n - 1.  CASE blocks fall through
unless BREAK is used, like in C.

LOOP takes an integer n (where n > 0) and repeats its body n times.
It's used to replace many repeated lines in the bigroom variants.

The OBJECT statement accepts new optional parameters:

 * buried
 * lit
 * eroded:n (-1 is erodeproof, 1-15 is erosion level)
 * locked/broken (for containers, broken means broken lock)
 * trapped (for containers)
 * recharged:n (the number of times recharged)
 * invisible
 * greased

The MONSTER statement accepts new optional parameters:

 * female
 * cancelled
 * revived (reduces dropped items on death)
 * avenge (pre-angered/safe to hit I think?)
 * fleeing:n (for 0 <= n <= 127 turns)
 * blinded:n (for 0 <= n <= 127 turns)
 * stunned
 * confused
 * seen_traps:x (x is all or pipe-delimited trap names mon has seen)

Compared to the UnNetHack changeset, buried objects on flipped levels
were fixed in a prior commit, and SPO_JE is fixed to use == for c
instead of = (masked by extra parentheses in UnNetHack); the latter is a
bug in UnNetHack, but the code there is replaced entirely in its most
recent version so there's no point reporting it.

According to UnNetHack, this is from SporkHack r531-538.