CS447: lev comp: new stormy level flag (Spork)
Adding this flag to the FLAGS of a level creates the thunder and
lightning effects of the Plane of Air, i.e. clouds shoot lightning and
cause thunder 1 in every 8 turns, paralyzing for 3 turns if in a cloud.

Compared to UnNetHack, this doesn't explicitly zero the stormy level
flag, since alloc_level() already zeroes out the whole level struct.  If
level struct fields need to be initialized non-zero, it should go there.

According to UnNetHack, this is from SporkHack r545.