CS448: lev comp: unhardcode quest mon gen (Spork)
This adds the new MON_GENERATION statement to level descriptions, for
use not only in quests but any level, like so:

    MON_GENERATION: 86%, (80,'S'), (16,"human mummy"), (4,'M')

The percentage is the chance of this custom spec overriding normal
monster generation.  The rest of the line consists of pairs of monsters
(or monster classes) in the ratio (not necessarily percentage) that they
will be randomly chosen when this override takes place.

Compared to UnNetHack:

 * The mon_gen_override and mon_gen_tuple structs are saved and restored
   by explicit formats, like the rest of NitroHack's save files.  In
   particular, this changeset's code was using old pointers saved with
   the structs to control the logic of restoring them, which caused me
   headaches.

 * rndmonst() (and rndmonnum() in turn) was changed to accept a full
   level up from just a dlevel, since the changes in it now pull out and
   use monster generation override info from it.  I chose this instead
   of a new argument because in practice all the callers of rndmonst()
   would use the same level for both.

According to UnNetHack, this is from SporkHack r546-548, r550.