CS450a: lev comp: Add monster inventories (Spork)
A monster's inventory can be set just like container contents, and so
the inventories of Pestilence, Orcus, Asmodeus, Croesus, Vlad the
Impaler, Dispater, Yeenoghu and Cthulhu are no longer hard-coded.

Compared to UnNetHack, this gives Croesus a two-handed sword in all 3
Fort Ludios variants, and removes the inventory hard-coding of Dispater,
Yeenoghu and Cthulhu along with the monsters already dealt with.

According to UnNetHack, this is from SporkHack r552 and r553.