CS453: lev comp: New level SOUNDS command (Spork)
This allows level-specific sounds to be defined for special levels:

    SOUNDS: 500, (hear, "a slithering sound."),
                 (feel, " a breeze."),
                 (verbal, "You shall both win and lose!"),
                 (pline, "You smell rotten eggs.")

The one of the above lines will be randomly chosen to be displayed one
in every 500 turns while on the level.

Compared to UnNetHack, this splits saving and freeing of level sounds
like the monster generation override commit, null-terminates restored
sound strings in case of random input (leaving out the null complicates
the save code), and removes the debugging pline() calls in
spo_level_sounds().

According to UnNetHack, this is from SporkHack r568.