CS457: lev comp:fix current_ hacks, 64-bit (Spork)
UnNetHack says this fixes an obscure 64-bit bug that caused a crash
while creating the Sanctum by replacing long with int; this is merged
with a similar fix for add_opvars() in case sizeof int != sizeof long,
so I can't tell where it is.  I don't even have a 64-bit CPU, so I'll
have to take these fixes on faith.

The current_region/current_coord hacks in the level compiler grammar
have been replaced with the use of Yacc features.

This changeset originally included a fix for ROOM/SUBROOM emitting
opcodes for IF statements, but I already rolled that into commit
6fc8caae1a (lev comp: value stack, chances, etc.).

According to UnNetHack, this is from SporkHack r590-592; r588 was
mentioned, but it's the ROOM/SUBROOM IF opcode fix.