CS490,509,526: Reduce Sokoban by one floor
Sokoban is now only 3 floors instead of 4.  The map variants of the
first and second floors have been merged into a single pool of 16
possible levels.

Changeset 490 does what was just described.  509 fixes a bug with
falling rock traps at the lowest levels of Sokoban.  526 adds the
scrolls of earth to the new bottom levels and places the downstairs for
them correctly.