Fix security bug in options handler (reported by Matthew Daley, fixed by Pasi Kallinen)
Quoting Matthew Daley:
Note the function 'escapes' in options.c, namely the branches starting
on lines 660 and 673. They process backslashes/carets in an escaped
string, converting them to the characters they represent. However,
they do not check that the character in the escaped string following
the backslash/caret is a valid one, and hence it can be anything,
including the null character. In this case the null character gets
written into the output buffer, and processing continues, skipping it.

This can be used to overflow the output buffer 'op_buf' on line 2108
in 'parseoptions' which is used to parse the 'menu_*' options and has
size BUFSZ, whereas the input buffer 'op' can be larger. When options
are being read from a config file through 'read_config_file' in
files.c, lines are read into a buffer of size BUFSZ*4 (line 1995). The
size of each comma-delimited option is then limited to only BUFSZ/2 by
'parseoptions' ('options.c' line 984), but this length check can be
bypassed by letting the faulty 'escapes' function continue parsing
over the null terminator (which was previously a comma separating
multiple options on a single config line) at the end of an option, and
onto the next one.


git-svn-id: https://svn.code.sf.net/p/unnethack/code/trunk@1048 f1e491ac-5ded-4967-80bf-9f4855be876a