Fix freeze when reviving Cthulhu on a full level
This fixes an infinite loop in makemon() when it's asked to randomly
place a monster (i.e. it's given coords 0,0) by giving up after 100
tries (arbitrary) to find a good position.  If it doesn't and
MM_ADJACENTOK is passed to makemon(), it tries to find an adjacent
position, and if that fails or the flag wasn't sent in, makemon()
returns NULL.

If Cthulhu tries to revive on a filled level, you'll "feel less
hassled", same as with the Riders.

This changes Cthulhu's revival slightly: if Cthulhu can't be placed in
the center of his cloud when the time comes, he'll be placed as close as
possible to it, instead of being placed randomly.  This removes
redundant placement code in revive_cthulhu() that really belongs in
makemon().

This is meant to go with e3fa9d98fe (CS140,151 add Cthulhu from SLASH'EM
Lethe).  This fixes #9.  This bug also exists in UnNetHack.  Credit to
Ion Frigate for discovering the bug.