Put dumpfiles in ~/.config/NitroHack/dumps, not ~
Right now, the change is only for Linux (or any system with UNIX
defined), but it wouldn't be hard to do the same for Windows, what with
its code being right below.

Also, compiling errors out early if not on Windows or Linux (really any
system with UNIX defined).

Based on the same change for NetHack 4 by Alex Smith:

https://gitorious.org/nitrohack/ais523/commit/9f9b7538413313cc2011e2d53c254f1eaf5a5a70