CS530: Fix shopkeeper msg on death when not seen
This fixes a logic error where a misplaced parenthesis sends a boolean
to cansee() where a y-coordinate is expected, causing the wrong
shopkeeper to be noted in the death message in a store when there are
multiple shopkeepers on the level.

Compared to UnNetHack, this also deletes some trailing whitespace on a
nearby line.