CS547: Allow engraving Elbereth with Ctrl-E
This implementation is based on a similar commit by Alex Smith from
NetHack4, though there's no reason the code from UnNetHack r547 couldn't
have been adapted:

https://gitorious.org/nitrohack/ais523/commit/24e00710cbd88e6630d0e9141ffc18bb3921f530

Compared to NetHack4, this actually returns the result of
doengrave_core() from doengrave() and doelbereth() (caught with compiler
warnings).  It also uses an athame if it's wielded (from UnNetHack
r547), and accounts for roleplaying illiterate conduct, which exists in
UnNetHack but not NetHack4.