CS565: Add giant turtle, vision-blocking monster
From SporkHack.  Has a field-of-view bug where zapping a giant turtle
with a wand of make invisible and walking around it for a few turns will
block vision, but displaced a couple of spaces to the right; it's
unlikely to be fixed, given how complex the NetHack vision code is.
Visible giant turtles are unaffected.