CS572,573: Allow levelport beyond Dungeons of Doom
Normally, a (non-wizard mode) player who chooses to level teleport from
the main dungeon to a level below the Valley of the Dead will experience
a random level teleport.  This change instead just puts them in the
Valley of the Dead.

If not level teleporting from in the main Dungeons of Doom branch, the
usual "Oops..." is replaced with "You feel like bouncing off a solid
wall!"