Fix saving/restoring player vampire character
Actually, this affects all characters, but the problem was introduced
with the vampire starting race commit.

Saving and restoring used numeric tags when handling the monster
associated with the player, which was a copy of one of the role monsters
towards the end of mons[] in monst.c, e.g. mons[PM_BARBARIAN].
NitroHack used pm_you_(fe)male to hold a mutable copy to mimic NetHack's
infravision hack, i.e. set the infravision flag on their role monster if
the player's race has infravision.

The vampire starting race change introduced upermonst, which was also a
mutable copy of the player's role monster, but substituted ALL of the
assumed-human flags with the flags of the player's race instead,
including infravision.  It turned out that pm_you_(fe)male was copied
into youmonst (used by the game), but upermonst was copied in right
after under exactly the same condition!  (See the definition for the
Upolyd macro and this commit's change to set_uasmon() to see how I
overlooked it.)

This fully removes pm_you_(fe)male and replaces its instances with
upermonst (esp. in the saving/restoring code, fixing a crash during
saving any character), since the latter does everything the former did.

This is meant to go with commit 8e01be9e0d (CS503,508,510 add vampire
player race... wow, a long time ago).