CS613,614: Replace poison instadeath w/ maxhp dmg
Instead of the instant death from poison without poison resistance ("The
poison was deadly..."), the player character now takes 4d6 HP worth of
extra damage and half of that to their maximum HP.

Compared to UnNetHack, this commit corrects the comment that says it
does damages "maxhp for (3d6)/2" when the code uses (4d6)/2.

According to UnNetHack, this is from SporkHack r136 and r137.