CS625: Add paranoid_trap option
When enabled (the default), players get a y/n prompt before walking onto
a trap.

Compared to UnNetHack, travelling/running causes the player to stop
before the trap instead of prompting, consistent with travelling/running
without the option, and the comment for the paranoid_trap flag doesn't
mention 'yes' because that isn't what's prompted for.