CS628: lev_comp:Give RANDOM_CORRIDORS optional arg
The RANDOM_CORRIDORS lev_comp statement can now take an optional
argument to specify 1 of 4 algorithms (3 new) to connect rooms and
corridors.

    RANDOM_CORRIDORS
    RANDOM_CORRIDORS: 0
        The default NetHack algorithm.

    RANDOM_CORRIDORS: 1
        Each room connects to a random other room.

    RANDOM_CORRIDORS: 2
        Circular connection: room1 -> room2 -> room3 -> ... -> room1

    RANDOM_CORRIDORS: 3
        Star connection: one room connects to all other rooms.

    RANDOM_CORRIDORS: random
        One of the above, chosen randomly.

The effect of the new algorithms is pretty subtle; because of the random
room placement and overlapping corridors, it's hard to tell if anything
has changed at all.  Even the most recent version of UnNetHack doesn't
use the new algorithms; SporkHack does in 10% of normal
rooms-and-corridors levels.

Compared to UnNetHack, this pulls in the lev_comp.y grammar changes from
changeset 630 that allow special level files to actually use this new
feature.

According to UnNetHack, this is from SporkHack r635.