Use new random corridor algorithms sometimes
10% of rooms-and-corridors levels will randomly choose from one of
random corridor algorithms, 3 out of 4 of which are new, so 7.5% of
levels will use one of the new ones.

I'm making this change even though it's not in UnNetHack because it's
such a shame to have this new feature and never have players experience
it.

From SporkHack; which revision?  Who knows; SporkHack's Subversion repo
is down again, so I can't get an annotate on the line this change comes
from.