Improve circular corridor algorithm
Levels with rooms connected using the new circular connection algorithm
now actually look circularly connected.  It works by sorting the rooms
by their rotational position relative to the center of the map.  Without
this, levels using the circular corridor algorithm looked like a mess of
corridors, with a long 'artery' corridor through the middle.

This uses a nested function and atan2() for sorting, which are C99, but
other places use snprintf(), so it's fine.  This also requires
libnitrohack to link against libm, C's standard math library.