CS629c,630a: lev_comp: add WALLWALK command
WALLWALK surrounds the inside of a normal rectangular room with tiles of
the level designer's choice, e.g. water or lava.

    WALLWALK: (5, 8), '}'
        Surround the room whose top-left room tile is at (5, 8) with
        moat water.

    WALLWALK: places[0], '}', 50%
        Same as above, but with a random location register and changing
        each tile only 50% of the time.

    WALLWALK: (5, 8), '}', 'L'
        Replace lava tiles with moat water.  Can be combined with the
        percentage chance above.

WALLWALK leaves traps and spaces next to doors alone.  When placing
lava, the lava spaces will always be lit.

WALLWALK replaces tiles downwards, then right, then up and finally left.
It stops and turns whenever it hits a tile that doesn't match its
expected background ('.' or room floor by default, lava in the 3rd
example above).  On unusual terrain, this might not come out as a
rectangle or surrounding ring, but maybe that's what you're after.

According to UnNetHack, this is one of a group of changes from SporkHack
r640-642.