CS639: lev_comp: Introduce $variables
This adds $variables to the special level compiler, which are created on
assignment, may be either integers or strings, and can be used in most
places where literal values are expected.

At least, that's what I think it does; there's no code whatsoever in
UnNetHack that actually runs this code, so I have no idea if it works,
or even if I've ported it correctly.  Guess we'll have to find out the
hard way...

According to UnNetHack, this is from SporkHack r666-667.