CS642: lev_comp: SporkHack r670-671
WALLWALK coordinate translation is fixed.

Map character variables and arrays are now supported by preceding a
character with 'TERRAIN:' when assigning such values.

WALLWALK can now accept map character variables.

Bugs with the length of OBJECT and MONSTER array variables have been
fixed.

Compared to UnNetHack, this removes a double-free from the SP_M_V_NAME
case of spo_monster(), which is also fixed in the latest version of
UnNetHack, but from which changeset I can't tell.

Since mapchar is actually used in the level compiler grammar, the number
of warnings from it is down from 14 to 10 with this change.