CS652: lev_comp: SporkHack r720,722,724,731,737-739,742
From the UnNetHack commit message:

 * Make get_location() handle irregular regions when a random coordinate
   is requested
 * Mess with the des-files[1]; randomize the amount of
   monsters/objects/traps, within sensible limits
 * Fix the impossible claiming we're comparing, when we're actually
   adding (ed: I fixed these error messages earlier when I added the
   code)
 * Add "selections" for TERRAIN and remove RANDLINE[2]
 * Add a new selection element floodfill(x,y) and a way to pick a random
   coordinate from a selection[3]
 * Add selection elements for circles and ellipses
 * Allow selection as DOOR position
 * New selection element "filter"[4]

[1] There are tons of little changes to many of the special levels,
particularly the quest levels, but also Mines and Gehennom.  It feels
like there are a few more monsters on average than anything else, but
not by too much.

[2] Level designers now have access to 'selections', which are sets of
coordinates on the map that can be manipulated, e.g. a randline
selection is made in the lava cavern levels of Gehennom to fill with
lava to form the rivers that cut the map in two.  That example is what
RANDLINE used to be for, so now it's gone.  Selections can also be
combined with logical OR using '&' (ampersand), or intersected with each
other with logical AND using 'filter(sel1, sel2)'.

[3] A floodfill terrain selection is used to find the coordinates of all
the street spaces in the Rogue quest start level to use as random
coordinates for monsters, for example.

[4] Intersects regions with logical AND; see [2].