CS654: lev_comp: SporkHack r762,765,771,779,780,785,791
From the UnNetHack commit message:

 * Fix some split-room buglets (ed: not really, see [1])
 * Make sure the Monk and Priest quest portals have a clear place to be
 * Allow selections as coordinates for FOUNTAIN, SINK and POOL[2]. Mess
   with bigrooms. (ed: slightly randomized terrain for some bigrooms)
 * Fix a crashing bug when creating sinks, fountains or pools (due to
   trying to free null spovar) (ed: probably introduced with these
   revisions too, since I see no trace such a problem nor a fix)
 * Handle the "transparent" mapchar (ed: use 'x' in maps to mark spaces
   to not draw over on the existing map)
 * Handle the "invisible" mapchar (ed: How is "invisible" different from
   "transparent"? How is this different from the above?)
 * Don't flip the levels without fixed map parts[3]
 * Fix randomly generated mine-type levels having lit walls when it
   shouldn't[4]
 * Don't complain when trying to give inventory to non-existent monster;
   the monster may be gone legally

[1] The split room port is practically all original code.  If the code
from this changeset were used, this would fix split rooms always being
at the top or left of the level, which is already fixed in the new split
room code.  The changeset also moves the add_room() call for the second
half of the split room from before to after a topologize() call, but in
writing original working split room code, that call is already done by
the usual rooms-and-corridors level generator, so in practice it would
do nothing.

[2] create_feature() is removed, since the same thing can be done
internally by replacing the terrain of a single tile selection.

[3] This is a workaround for the ugly left-most walls in flipped
mines-like levels.  Compared to UnNetHack, Cav-fila gets this fix too.

[4] This adds a missing argument to lvlfill_solid() for the
LVLINIT_MINES case... one that has been missing for over 200 changesets
(since changeset 442)!  How this didn't break compilation for UnNetHack
and remain undetected for so long is a mystery to me...